First Gate Function

This function checks what are available nodes to split and then call find_split() on each node, then decide which node creates best split, and call splitter() to perform the split.

Usage

checkem(
  toclust.fd,
  frame,
  cloc,
  dist,
  dsubs,
  dsubsname,
  weights,
  minbucket,
  minsplit,
  spliton,
  method
)

Arguments

toclust.fd

A functional data object (i.e., having class fd) created from fda package. See fda::fd().

frame

The split tree transferred as data frame.

cloc

Vector of current cluster membership.

dist

Distance matrix of all observations in the data.

dsubs

Distance matrix calculated on each subregion. A three-dimensional matrix.

dsubsname

Subregion names.

weights

(Currently unused) Weights on observations.

minbucket

The minimum number of data points in one cluster allowed.

minsplit

The minimum number of observations that must exist in a node in order for a split to be attempted.

spliton

Restrict the partitioning on a specific set of subregions.

method

The clustering method you want to run in each subregion. Can be chosen between pam and ward.

Value

It is not supposed to return anything because global environment was used. However, if there is nothing left to split, it returns 0 to tell the caller to stop running the loop.