Find the Best Split

Find the best split in terms of reduction in inertia for the transferred node, indicate by row. Find the terminal node with the greatest change in inertia and bi-partition it.

Usage

find_split(
  toclust.fd,
  frame_row,
  cloc,
  dist,
  dsubs,
  dsubsname,
  weights,
  minbucket,
  minsplit,
  spliton,
  method
)

Arguments

toclust.fd

A functional data object (i.e., having class fd) created from fda package. See fda::fd().

frame_row

One row of the split tree as data frame.

cloc

Vector of current cluster membership.

dist

Distance matrix of all observations in the data.

dsubs

Distance matrix calculated on each subregion. A three-dimensional matrix.

dsubsname

Subregion names.

weights

(Currently unused) Weights on observations.

minbucket

The minimum number of data points in one cluster allowed.

minsplit

The minimum number of observations that must exist in a node in order for a split to be attempted.

spliton

Restrict the partitioning on a specific set of subregions.

method

The clustering method you want to run in each subregion. Can be chosen between pam and ward.

Value

The updated frame_row with the next split updated.