Split Function

Given the Cluster's frame's row position to split at split_row, this function performs the split, calculate all necessary information for the splitting tree and cluster memberships.

Usage

splitter(
  toclust.fd,
  split_row,
  frame,
  cloc,
  dist,
  dsubs,
  dsubsname,
  weights,
  method
)

Arguments

toclust.fd

A functional data object (i.e., having class fd) created from fda package. See fda::fd().

split_row

The row index in frame that would be split on.

frame

The split tree transferred as data frame.

cloc

Vector of current cluster membership.

dist

Distance matrix of all observations in the data.

dsubs

Distance matrix calculated on each subregion. A three-dimensional matrix.

dsubsname

Subregion names.

weights

(Currently unused) Weights on observations.

method

The clustering method you want to run in each subregion. Can be chosen between pam and ward.

Value

Updated frame and cloc saved in a list.