Distance Between Functional Objects

Calculate the distance between functional objects over the defined range.

fdistmatrix(fd, subrange, distmethod)

Arguments

fd
A functional data object fd of fda package.

subrange
A vector of two values indicating the value range of functional object to calculate on.

distmethod
The method for calculating the distance matrix. Choose between "usc" and "manual". "usc" uses fda.usc::metric.lp() function while "manual" uses squared distance between functions. See @details.

Value

A distance matrix with diagonal value and the upper half.

Details

If choosing distmethod = "manual", the L2 distance between all pairs of functions \(y_i(t)\) and \(y_j(t)\) is given by: $$d_R(y_i, y_j) = \sqrt{\int_{a_r}^{b_r} [y_i(t) - y_j(t)]^2 dt}.$$

Examples

library(fda)

#> Loading required package: Matrix

#> 
#> Attaching package: 'fda'

#> The following object is masked from 'package:graphics':
#> 
#>     matplot

# Examples taken from fda::Data2fd()
data(gait)
# Function only works on two dimensional data
gait <- gait[, , 1]
gaitbasis3 <- create.fourier.basis(nbasis = 5)
gaitfd3 <- Data2fd(gait, basisobj = gaitbasis3)

#> 'y' is missing, using 'argvals'
#> 'argvals' is missing;  using seq( 0 ,  1 , length= 20 )

fdistmatrix(gaitfd3, c(0.2, 0.4), "usc")

#> Warning: executing %dopar% sequentially: no parallel backend registered

#>            boy1      boy2      boy3      boy4      boy5      boy6      boy7
#> boy1  0.0000000 1.5087312 2.7026782 2.9596849  6.555144 2.7342900 1.3753866
#> boy2  1.5087312 0.0000000 2.0897922 3.7401208  7.326593 2.6888444 2.4864545
#> boy3  2.7026782 2.0897922 0.0000000 5.5707196  9.180870 1.1719472 4.0605859
#> boy4  2.9596849 3.7401208 5.5707196 0.0000000  3.643980 5.6727584 1.6234050
#> boy5  6.5551441 7.3265931 9.1808698 3.6439803  0.000000 9.2665974 5.1945234
#> boy6  2.7342900 2.6888444 1.1719472 5.6727584  9.266597 0.0000000 4.0743911
#> boy7  1.3753866 2.4864545 4.0605859 1.6234050  5.194523 4.0743911 0.0000000
#> boy8  4.3551797 3.7580961 1.7142706 7.2659453 10.880113 1.9294501 5.7271108
#> boy9  2.3164751 2.4162985 1.1955910 5.2653932  8.843621 0.4692383 3.6579016
#> boy10 2.2404441 2.7053925 4.6472878 1.1281579  4.625525 4.9084048 1.2419640
#> boy11 1.1059681 0.9798274 2.9049237 2.7878192  6.366207 3.3183554 1.6104968
#> boy12 0.5008053 1.5615395 3.0512714 2.5502560  6.143839 3.1957811 1.0451135
#> boy13 1.5302662 2.6359376 4.2075153 1.5079093  5.034133 4.2393033 0.2531336
#> boy14 0.7233230 1.0167148 2.0592524 3.5490611  7.178694 2.2092448 2.0300579
#> boy15 1.3206655 1.8762233 1.7489368 4.2589830  7.820273 1.4773117 2.6446240
#> boy16 1.0129111 0.7916325 1.9311996 3.6704324  7.259722 2.3523503 2.2260119
#> boy17 1.9024199 0.4814183 1.7897449 4.2109209  7.791263 2.5200239 2.9549100
#> boy18 1.6959680 0.8674292 1.2877911 4.3857076  8.013177 1.8456934 2.9569147
#> boy19 2.4573259 3.2414307 5.0457535 0.5896231  4.137555 5.1877289 1.1703665
#> boy20 2.2299524 2.7306678 4.6637632 1.0713446  4.604612 4.9004000 1.1923201
#> boy21 3.5555109 4.2119214 6.1108007 0.6588985  3.123284 6.2745359 2.2491793
#> boy22 1.4631030 1.8960782 3.8046773 1.8524460  5.463941 4.0479390 0.9735384
#> boy23 4.9275841 4.6055451 2.5617875 7.8797465 11.475173 2.2251688 6.2855212
#> boy24 1.6660422 2.5456632 4.2784994 1.3043997  4.916384 4.3923115 0.4564002
#> boy25 0.6678486 1.1881653 2.8544507 2.7270052  6.331978 3.1268877 1.3432015
#> boy26 2.4573259 3.2414307 5.0457535 0.5896231  4.137555 5.1877289 1.1703665
#> boy27 3.3248524 3.4457901 5.4800418 1.6180047  4.219183 5.9023320 2.4125651
#> boy28 1.4159892 1.9207285 3.8009651 1.8374126  5.462194 4.0151710 0.8986901
#> boy29 3.1716664 3.8755797 5.7451426 0.3087153  3.463242 5.8929929 1.8613630
#> boy30 2.4538938 1.5386473 0.6550832 5.1792040  8.787188 1.6639955 3.7506011
#> boy31 2.7342032 2.4500569 0.8441356 5.6562970  9.278966 0.4530506 4.0900264
#> boy32 3.6644009 3.8193263 2.1362432 6.5659488 10.083299 1.1534448 4.9463599
#> boy33 3.1315183 3.5150495 5.4943088 1.0321086  3.896961 5.8065344 2.0393326
#> boy34 1.2694508 1.2495416 3.2055680 2.5266438  6.085741 3.6097040 1.4756647
#> boy35 2.3286742 1.8793554 0.4300901 5.2304590  8.847764 0.9771081 3.6956142
#> boy36 2.0235868 1.5476128 0.6846785 4.8999393  8.516565 1.2327368 3.3809718
#> boy37 1.5285074 1.3410014 3.3630611 2.5025059  6.017423 3.8152269 1.6197263
#> boy38 7.0291617 6.6423223 4.5830699 9.9825469 13.573942 4.3223450 8.3866192
#> boy39 3.6330940 2.8262831 1.0005962 6.4496901 10.074811 1.6712405 4.9750004
#>            boy8      boy9     boy10     boy11     boy12     boy13     boy14
#> boy1   4.355180 2.3164751 2.2404441 1.1059681 0.5008053 1.5302662 0.7233230
#> boy2   3.758096 2.4162985 2.7053925 0.9798274 1.5615395 2.6359376 1.0167148
#> boy3   1.714271 1.1955910 4.6472878 2.9049237 3.0512714 4.2075153 2.0592524
#> boy4   7.265945 5.2653932 1.1281579 2.7878192 2.5502560 1.5079093 3.5490611
#> boy5  10.880113 8.8436206 4.6255253 6.3662067 6.1438393 5.0341330 7.1786945
#> boy6   1.929450 0.4692383 4.9084048 3.3183554 3.1957811 4.2393033 2.2092448
#> boy7   5.727111 3.6579016 1.2419640 1.6104968 1.0451135 0.2531336 2.0300579
#> boy8   0.000000 2.2829493 6.3561522 4.6135714 4.7382094 5.8779312 3.7255759
#> boy9   2.282949 0.0000000 4.5179850 2.9580584 2.7783489 3.8127658 1.8398423
#> boy10  6.356152 4.5179850 0.0000000 1.7645559 1.7788012 1.2036625 2.7074621
#> boy11  4.613571 2.9580584 1.7645559 0.0000000 0.8446140 1.7292912 1.1671847
#> boy12  4.738209 2.7783489 1.7788012 0.8446140 0.0000000 1.1697540 1.0783269
#> boy13  5.877931 3.8127658 1.2036625 1.7292912 1.1697540 0.0000000 2.2012305
#> boy14  3.725576 1.8398423 2.7074621 1.1671847 1.0783269 2.2012305 0.0000000
#> boy15  3.216479 1.0289808 3.5547751 2.1258753 1.7916613 2.7906735 1.0197131
#> boy16  3.642624 1.9831307 2.7450157 1.0537186 1.1885534 2.3521883 0.5996949
#> boy17  3.394526 2.3169159 3.1675346 1.4579499 2.0155859 3.1093396 1.3008233
#> boy18  2.955485 1.6242825 3.4297304 1.6982447 1.9717991 3.1227714 0.9885366
#> boy19  6.746935 4.7719377 0.7415961 2.2721631 2.0217325 1.0278442 3.0520104
#> boy20  6.370377 4.5116341 0.1224981 1.7942192 1.7795794 1.1653132 2.7053112
#> boy21  7.817182 5.8684851 1.5131084 3.2629796 3.1207300 2.1226530 4.1205455
#> boy22  5.504737 3.6737049 0.8899362 0.9872120 1.0530236 1.0922139 1.8480275
#> boy23  1.157679 2.6367036 7.0655131 5.3859530 5.3661432 6.4414426 4.3634247
#> boy24  5.971146 3.9804266 0.7966981 1.5995721 1.2467893 0.4203548 2.2635301
#> boy25  4.561403 2.7334782 1.8228806 0.4638374 0.4050837 1.4712567 0.9435785
#> boy26  6.746935 4.7719377 0.7415961 2.2721631 2.0217325 1.0278442 3.0520104
#> boy27  7.179358 5.5262804 1.1917502 2.6033071 2.8440768 2.3356311 3.7031694
#> boy28  5.497555 3.6394775 0.9225993 1.0214588 1.0177300 1.0325588 1.8214577
#> boy29  7.448992 5.4847550 1.1975841 2.9196699 2.7409026 1.7346520 3.7468412
#> boy30  2.230820 1.5959600 4.1966125 2.4357039 2.7336277 3.9006706 1.7698782
#> boy31  1.757890 0.7142540 4.8283613 3.1768174 3.1682073 4.2607820 2.1271448
#> boy32  2.037891 1.4275703 5.8950078 4.3821859 4.1425330 5.0959663 3.2489938
#> boy33  7.205953 5.4115491 0.9160626 2.5946994 2.6546120 1.9346235 3.6033669
#> boy34  4.914575 3.2451687 1.4796548 0.3104749 0.9106465 1.5780324 1.4397493
#> boy35  2.036404 0.8826883 4.3424925 2.6313643 2.7067709 3.8494027 1.6925477
#> boy36  2.367634 1.0304363 3.9996348 2.2832587 2.3810424 3.5331194 1.3764768
#> boy37  5.063515 3.4630503 1.4079690 0.5103191 1.1633737 1.7097486 1.6631132
#> boy38  2.930183 4.7327561 9.1590610 7.4600771 7.4647555 8.5404104 6.4615807
#> boy39  1.035900 1.9319361 5.4902220 3.7384832 3.9763146 5.1355106 2.9489503
#>          boy15     boy16     boy17     boy18     boy19     boy20      boy21
#> boy1  1.320665 1.0129111 1.9024199 1.6959680 2.4573259 2.2299524  3.5555109
#> boy2  1.876223 0.7916325 0.4814183 0.8674292 3.2414307 2.7306678  4.2119214
#> boy3  1.748937 1.9311996 1.7897449 1.2877911 5.0457535 4.6637632  6.1108007
#> boy4  4.258983 3.6704324 4.2109209 4.3857076 0.5896231 1.0713446  0.6588985
#> boy5  7.820273 7.2597225 7.7912630 8.0131771 4.1375551 4.6046124  3.1232840
#> boy6  1.477312 2.3523503 2.5200239 1.8456934 5.1877289 4.9004000  6.2745359
#> boy7  2.644624 2.2260119 2.9549100 2.9569147 1.1703665 1.1923201  2.2491793
#> boy8  3.216479 3.6426240 3.3945263 2.9554845 6.7469353 6.3703766  7.8171817
#> boy9  1.028981 1.9831307 2.3169159 1.6242825 4.7719377 4.5116341  5.8684851
#> boy10 3.554775 2.7450157 3.1675346 3.4297304 0.7415961 0.1224981  1.5131084
#> boy11 2.125875 1.0537186 1.4579499 1.6982447 2.2721631 1.7942192  3.2629796
#> boy12 1.791661 1.1885534 2.0155859 1.9717991 2.0217325 1.7795794  3.1207300
#> boy13 2.790673 2.3521883 3.1093396 3.1227714 1.0278442 1.1653132  2.1226530
#> boy14 1.019713 0.5996949 1.3008233 0.9885366 3.0520104 2.7053112  4.1205455
#> boy15 0.000000 1.2640655 1.9860227 1.3761413 3.7623700 3.5463591  4.8658232
#> boy16 1.264065 0.0000000 1.0708765 0.9138141 3.1302160 2.7662302  4.1951278
#> boy17 1.986023 1.0708765 0.0000000 0.6980734 3.7173842 3.1929973  4.6742290
#> boy18 1.376141 0.9138141 0.6980734 0.0000000 3.8836286 3.4420524  4.9109932
#> boy19 3.762370 3.1302160 3.7173842 3.8836286 0.0000000 0.7046188  1.1179242
#> boy20 3.546359 2.7662302 3.1929973 3.4420524 0.7046188 0.0000000  1.4849580
#> boy21 4.865823 4.1951278 4.6742290 4.9109932 1.1179242 1.4849580  0.0000000
#> boy22 2.743907 1.9406949 2.3627384 2.5718816 1.3992827 0.8807673  2.3491499
#> boy23 3.660887 4.3861882 4.2938442 3.7511964 7.3801239 7.0667405  8.4698540
#> boy24 2.976429 2.3852360 3.0234284 3.1259235 0.7987671 0.7546364  1.8904000
#> boy25 1.815539 0.9483449 1.6561872 1.7080583 2.1971055 1.8369402  3.2596939
#> boy26 3.762370 3.1302160 3.7173842 3.8836286 0.0000000 0.7046188  1.1179242
#> boy27 4.604051 3.6242061 3.8575940 4.2678183 1.5391003 1.2571508  1.4748998
#> boy28 2.703000 1.9428509 2.3879768 2.5717992 1.3901070 0.9035504  2.3540135
#> boy29 4.479882 3.8310127 4.3430750 4.5545331 0.7349495 1.1612888  0.3888333
#> boy30 1.812010 1.5763248 1.1794876 0.8367913 4.6602957 4.2188279  5.6891698
#> boy31 1.583899 2.2284519 2.2209053 1.5847740 5.1678908 4.8258286  6.2409232
#> boy32 2.351534 3.4030088 3.6631748 2.9875465 6.0918580 5.8823671  7.1889364
#> boy33 4.443907 3.5884506 3.9609583 4.2846326 1.0143513 0.9556591  0.9486709
#> boy34 2.386859 1.3436544 1.7184080 1.9954122 2.0156580 1.5129637  2.9793876
#> boy35 1.340846 1.6463422 1.6507311 1.0371321 4.7146145 4.3524061  5.7878685
#> boy36 1.176768 1.3008027 1.3581393 0.7259875 4.3810043 4.0111769  5.4503175
#> boy37 2.629772 1.5480399 1.7886993 2.1422455 2.0143432 1.4521595  2.9131539
#> boy38 5.755293 6.4609117 6.2934151 5.8035775 9.4807722 9.1622042 10.5710954
#> boy39 2.678291 2.8480660 2.4263705 2.0683186 5.9406964 5.5066226  6.9771147
#>           boy22     boy23     boy24     boy25     boy26      boy27     boy28
#> boy1  1.4631030  4.927584 1.6660422 0.6678486 2.4573259  3.3248524 1.4159892
#> boy2  1.8960782  4.605545 2.5456632 1.1881653 3.2414307  3.4457901 1.9207285
#> boy3  3.8046773  2.561787 4.2784994 2.8544507 5.0457535  5.4800418 3.8009651
#> boy4  1.8524460  7.879746 1.3043997 2.7270052 0.5896231  1.6180047 1.8374126
#> boy5  5.4639414 11.475173 4.9163841 6.3319780 4.1375551  4.2191829 5.4621941
#> boy6  4.0479390  2.225169 4.3923115 3.1268877 5.1877289  5.9023320 4.0151710
#> boy7  0.9735384  6.285521 0.4564002 1.3432015 1.1703665  2.4125651 0.8986901
#> boy8  5.5047375  1.157679 5.9711464 4.5614032 6.7469353  7.1793579 5.4975549
#> boy9  3.6737049  2.636704 3.9804266 2.7334782 4.7719377  5.5262804 3.6394775
#> boy10 0.8899362  7.065513 0.7966981 1.8228806 0.7415961  1.1917502 0.9225993
#> boy11 0.9872120  5.385953 1.5995721 0.4638374 2.2721631  2.6033071 1.0214588
#> boy12 1.0530236  5.366143 1.2467893 0.4050837 2.0217325  2.8440768 1.0177300
#> boy13 1.0922139  6.441443 0.4203548 1.4712567 1.0278442  2.3356311 1.0325588
#> boy14 1.8480275  4.363425 2.2635301 0.9435785 3.0520104  3.7031694 1.8214577
#> boy15 2.7439070  3.660887 2.9764294 1.8155393 3.7623700  4.6040511 2.7029999
#> boy16 1.9406949  4.386188 2.3852360 0.9483449 3.1302160  3.6242061 1.9428509
#> boy17 2.3627384  4.293844 3.0234284 1.6561872 3.7173842  3.8575940 2.3879768
#> boy18 2.5718816  3.751196 3.1259235 1.7080583 3.8836286  4.2678183 2.5717992
#> boy19 1.3992827  7.380124 0.7987671 2.1971055 0.0000000  1.5391003 1.3901070
#> boy20 0.8807673  7.066741 0.7546364 1.8369402 0.7046188  1.2571508 0.9035504
#> boy21 2.3491499  8.469854 1.8904000 3.2596939 1.1179242  1.4748998 2.3540135
#> boy22 0.0000000  6.202758 0.7930559 1.0268006 1.3992827  1.9383615 0.1029022
#> boy23 6.2027585  0.000000 6.5886912 5.2575853 7.3801239  7.9871824 6.1799051
#> boy24 0.7930559  6.588691 0.0000000 1.4499951 0.7987671  1.9579587 0.7500925
#> boy25 1.0268006  5.257585 1.4499951 0.0000000 2.1971055  2.7988152 1.0214256
#> boy26 1.3992827  7.380124 0.7987671 2.1971055 0.0000000  1.5391003 1.3901070
#> boy27 1.9383615  7.987182 1.9579587 2.7988152 1.5391003  0.0000000 2.0043889
#> boy28 0.1029022  6.179905 0.7500925 1.0214256 1.3901070  2.0043889 0.0000000
#> boy29 2.0032554  8.090411 1.5057750 2.8930347 0.7349495  1.4375575 2.0023593
#> boy30 3.3636272  3.156815 3.9144208 2.4709796 4.6602957  4.9623530 3.3712505
#> boy31 3.9564663  2.259151 4.3734115 3.0445421 5.1678908  5.7740226 3.9318195
#> boy32 5.0608023  1.701693 5.3073693 4.1380081 6.0918580  6.9349425 5.0188151
#> boy33 1.7934867  7.947486 1.5909844 2.6940961 1.0143513  0.6039656 1.8342939
#> boy34 0.7525528  5.687156 1.3918524 0.6184472 2.0156580  2.3027256 0.8082721
#> boy35 3.4890340  2.759074 3.9356361 2.5377057 4.7146145  5.2299485 3.4772607
#> boy36 3.1497346  3.108323 3.6065941 2.1972026 4.3810043  4.8807384 3.1406401
#> boy37 0.7788626  5.865576 1.4678516 0.8784564 2.0143432  2.1270481 0.8562766
#> boy38 8.2956372  2.103007 8.6905178 7.3492616 9.4807722 10.0541010 8.2750992
#> boy39 4.6390835  2.071556 5.1769196 3.7479346 5.9406964  6.2650556 4.6394832
#>            boy29     boy30     boy31     boy32      boy33     boy34     boy35
#> boy1   3.1716664 2.4538938 2.7342032  3.664401  3.1315183 1.2694508 2.3286742
#> boy2   3.8755797 1.5386473 2.4500569  3.819326  3.5150495 1.2495416 1.8793554
#> boy3   5.7451426 0.6550832 0.8441356  2.136243  5.4943088 3.2055680 0.4300901
#> boy4   0.3087153 5.1792040 5.6562970  6.565949  1.0321086 2.5266438 5.2304590
#> boy5   3.4632421 8.7871884 9.2789657 10.083299  3.8969613 6.0857411 8.8477644
#> boy6   5.8929929 1.6639955 0.4530506  1.153445  5.8065344 3.6097040 0.9771081
#> boy7   1.8613630 3.7506011 4.0900264  4.946360  2.0393326 1.4756647 3.6956142
#> boy8   7.4489916 2.2308203 1.7578900  2.037891  7.2059532 4.9145755 2.0364043
#> boy9   5.4847550 1.5959600 0.7142540  1.427570  5.4115491 3.2451687 0.8826883
#> boy10  1.1975841 4.1966125 4.8283613  5.895008  0.9160626 1.4796548 4.3424925
#> boy11  2.9196699 2.4357039 3.1768174  4.382186  2.5946994 0.3104749 2.6313643
#> boy12  2.7409026 2.7336277 3.1682073  4.142533  2.6546120 0.9106465 2.7067709
#> boy13  1.7346520 3.9006706 4.2607820  5.095966  1.9346235 1.5780324 3.8494027
#> boy14  3.7468412 1.7698782 2.1271448  3.248994  3.6033669 1.4397493 1.6925477
#> boy15  4.4798817 1.8120099 1.5838986  2.351534  4.4439068 2.3868585 1.3408463
#> boy16  3.8310127 1.5763248 2.2284519  3.403009  3.5884506 1.3436544 1.6463422
#> boy17  4.3430750 1.1794876 2.2209053  3.663175  3.9609583 1.7184080 1.6507311
#> boy18  4.5545331 0.8367913 1.5847740  2.987547  4.2846326 1.9954122 1.0371321
#> boy19  0.7349495 4.6602957 5.1678908  6.091858  1.0143513 2.0156580 4.7146145
#> boy20  1.1612888 4.2188279 4.8258286  5.882367  0.9556591 1.5129637 4.3524061
#> boy21  0.3888333 5.6891698 6.2409232  7.188936  0.9486709 2.9793876 5.7878685
#> boy22  2.0032554 3.3636272 3.9564663  5.060802  1.7934867 0.7525528 3.4890340
#> boy23  8.0904108 3.1568152 2.2591510  1.701693  7.9474863 5.6871558 2.7590743
#> boy24  1.5057750 3.9144208 4.3734115  5.307369  1.5909844 1.3918524 3.9356361
#> boy25  2.8930347 2.4709796 3.0445421  4.138008  2.6940961 0.6184472 2.5377057
#> boy26  0.7349495 4.6602957 5.1678908  6.091858  1.0143513 2.0156580 4.7146145
#> boy27  1.4375575 4.9623530 5.7740226  6.934943  0.6039656 2.3027256 5.2299485
#> boy28  2.0023593 3.3712505 3.9318195  5.018815  1.8342939 0.8082721 3.4772607
#> boy29  0.0000000 5.3362100 5.8652771  6.801963  0.8521891 2.6437757 5.4171172
#> boy30  5.3362100 0.0000000 1.3000477  2.730784  5.0215940 2.7305704 0.7213204
#> boy31  5.8652771 1.3000477 0.0000000  1.509611  5.7198913 3.4754853 0.7009013
#> boy32  6.8019630 2.7307836 1.5096114  0.000000  6.7914411 4.6668546 2.0590761
#> boy33  0.8521891 5.0215940 5.7198913  6.791441  0.0000000 2.2970701 5.2091052
#> boy34  2.6437757 2.7305704 3.4754853  4.666855  2.2970701 0.0000000 2.9335852
#> boy35  5.4171172 0.7213204 0.7009013  2.059076  5.2091052 2.9335852 0.0000000
#> boy36  5.0814646 0.6549597 0.9977465  2.337424  4.8625876 2.5848882 0.3503943
#> boy37  2.5967766 2.8558577 3.6593764  4.887926  2.1765565 0.2684835 3.1090216
#> boy38 10.1920863 5.1422452 4.3509019  3.628120 10.0352824 7.7617398 4.8295309
#> boy39  6.6200594 1.3161893 1.2998443  2.339394  6.3270862 4.0346228 1.3544788
#>           boy36     boy37     boy38     boy39
#> boy1  2.0235868 1.5285074  7.029162  3.633094
#> boy2  1.5476128 1.3410014  6.642322  2.826283
#> boy3  0.6846785 3.3630611  4.583070  1.000596
#> boy4  4.8999393 2.5025059  9.982547  6.449690
#> boy5  8.5165651 6.0174227 13.573942 10.074811
#> boy6  1.2327368 3.8152269  4.322345  1.671241
#> boy7  3.3809718 1.6197263  8.386619  4.975000
#> boy8  2.3676340 5.0635149  2.930183  1.035900
#> boy9  1.0304363 3.4630503  4.732756  1.931936
#> boy10 3.9996348 1.4079690  9.159061  5.490222
#> boy11 2.2832587 0.5103191  7.460077  3.738483
#> boy12 2.3810424 1.1633737  7.464755  3.976315
#> boy13 3.5331194 1.7097486  8.540410  5.135511
#> boy14 1.3764768 1.6631132  6.461581  2.948950
#> boy15 1.1767675 2.6297719  5.755293  2.678291
#> boy16 1.3008027 1.5480399  6.460912  2.848066
#> boy17 1.3581393 1.7886993  6.293415  2.426370
#> boy18 0.7259875 2.1422455  5.803578  2.068319
#> boy19 4.3810043 2.0143432  9.480772  5.940696
#> boy20 4.0111769 1.4521595  9.162204  5.506623
#> boy21 5.4503175 2.9131539 10.571095  6.977115
#> boy22 3.1497346 0.7788626  8.295637  4.639084
#> boy23 3.1083234 5.8655763  2.103007  2.071556
#> boy24 3.6065941 1.4678516  8.690518  5.176920
#> boy25 2.1972026 0.8784564  7.349262  3.747935
#> boy26 4.3810043 2.0143432  9.480772  5.940696
#> boy27 4.8807384 2.1270481 10.054101  6.265056
#> boy28 3.1406401 0.8562766  8.275099  4.639483
#> boy29 5.0814646 2.5967766 10.192086  6.620059
#> boy30 0.6549597 2.8558577  5.142245  1.316189
#> boy31 0.9977465 3.6593764  4.350902  1.299844
#> boy32 2.3374243 4.8879256  3.628120  2.339394
#> boy33 4.8625876 2.1765565 10.035282  6.327086
#> boy34 2.5848882 0.2684835  7.761740  4.034623
#> boy35 0.3503943 3.1090216  4.829531  1.354479
#> boy36 0.0000000 2.7610333  5.177935  1.624878
#> boy37 2.7610333 0.0000000  7.930361  4.159231
#> boy38 5.1779347 7.9303611  0.000000  3.914306
#> boy39 1.6248780 4.1592311  3.914306  0.000000